ScanSage_Frontend/ScanSage_UI/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.1.31
chainlit==2.4.201
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.18
fastapi==0.115.11
filetype==1.2.0
frozenlist==1.5.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
Lazify==0.4.0
literalai==0.1.103
marshmallow==3.26.1
mcp==1.5.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.41.3
syncer==2.0.3
tomli==2.2.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.31.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
